+++ lib/Vocoder.cc
@@ -14,15 +14,15 @@ namespace pwv {
 
 namespace {
 
-std::vector<float> bandpass(float sampling_rate, std::span<float const> input,
-                            float hz, float q) {
+std::vector<float> bandpass(double sampling_rate, std::span<float const> input,
+                            double hz, double q) {
     std::vector<float> result(input.begin(), input.end());
     BandPass(sampling_rate, hz, q).process(result);
     return result;
 }
 
-std::vector<float> lowpass(float sampling_rate, std::span<float const> input,
-                           float cutoff) {
+std::vector<float> lowpass(double sampling_rate, std::span<float const> input,
+                           double cutoff) {
     std::vector<float> result(input.begin(), input.end());
     LowPass(sampling_rate, cutoff).process(result);
     return result;
@@ -50,11 +50,11 @@ void add(std::span<float> a, std::span<float const> b) {
 
 }  // namespace
 
-Vocoder::Vocoder(int distance, int num_bands, int sampling_rate)
+Vocoder::Vocoder(double distance, int num_bands, double sampling_rate)
     : m_distance(distance),
       m_num_bands(num_bands),
       m_sampling_rate(sampling_rate),
-      m_octaves(std::log2(m_sampling_rate / 2.205f / 20)),
+      m_octaves(std::log2(m_sampling_rate / 2.205 / 20)),
       m_interval(12 * m_octaves / m_num_bands) {}
 
 Vocoder::~Vocoder() {}
@@ -64,8 +64,8 @@ std::vector<float> Vocoder::process(std::span<float const> input,
     assert(input.size() == carrier.size());
     std::vector<float> result(input.size());
 
-    float const q = m_octaves / m_num_bands / std::sqrt(2);
-    float band_hz = next_hz(20, m_interval / 2.f);
+    double const q = m_octaves / m_num_bands / std::sqrt(2);
+    double band_hz = next_hz(20, m_interval / 2.f);
     for (int band = 0; band < m_num_bands; band++) {
         // Bandpass both
         auto input_filtered = bandpass(m_sampling_rate, input, band_hz, q);
@@ -87,15 +87,15 @@ std::vector<float> Vocoder::process(std::span<float const> input,
     return result;
 }
 
-float Vocoder::next_hz(float hz, float interval) const {
+double Vocoder::next_hz(double hz, double interval) const {
     // TODO: optimize this
-    float const k = std::log(2) / 12;
-    float const c = std::log(440) - k * 69;
-    auto hz_to_step = [&](float pitch) { return (std::log(pitch) - c) / k; };
-    auto step_to_hz = [&](float step) { return std::exp(k * step + c); };
+    double const k = std::log(2) / 12;
+    double const c = std::log(440) - k * 69;
+    auto hz_to_step = [&](double pitch) { return (std::log(pitch) - c) / k; };
+    auto step_to_hz = [&](double step) { return std::exp(k * step + c); };
 
-    float const prev_step = hz_to_step(hz);
-    float const next_step = prev_step + interval;
+    double const prev_step = hz_to_step(hz);
+    double const next_step = prev_step + interval;
     return step_to_hz(next_step);
 }
